+++ plugins/modules/volume_info.py
@@ -319,8 +319,8 @@ def get_sdk_config(module, sdk):
 def check_required_arguments(module, object_name):
     # manually checking if token or username & password provided
     if (
-            not module.params.get("token")
-            and not (module.params.get("username") and module.params.get("password"))
+        not module.params.get("token")
+        and not (module.params.get("username") and module.params.get("password"))
     ):
         module.fail_json(
             msg='Token or username & password are required for {object_name}'.format(